Create T and rho pairs for I.C.
I ran through the temperature and density initial conditions and found which values give an immediate error small enough for us, so we can use these values when plotting to save time. At around T_c = 1e6, the error diverged, and also diverged for the expected T and rho that should work for a low mass star.